- get version from mpay999

+++ DeMobile/Models/AppModel/m_Customer.cs
@@ -21,7 +21,7 @@ namespace DeMobile.Models.AppModel
         public string CITIZEN_NO { get; set; }
         public string TEL { get; set; }
         public string PERMIT { get; set; }
-        public int APP_VERSION { get; set; }
+        public double APP_VERSION { get; set; }
         public string CHAT { get; set; }
     }
     public class m_device
